requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib
argon2-cffi
razorpay
requests
psycopg2-binary
boto3
python-dotenv
pydantic-settings
